framenum was an int, now int64_t and yes, I am aware client + server code needs 64 bits fixing too.

+++ src/baseq2/svgame/g_local.h
@@ -425,7 +425,7 @@ typedef struct {
 // it is read/written to the level.sav file for savegames
 //
 typedef struct {
-    int         framenum;
+    int64_t         framenum;
     sg_time_t		time;
 
     char        level_name[MAX_QPATH];  // the descriptive name (Outer Base, etc)
